disassemble: extract score and fields
