+++ strcs/disassemble/base.py
@@ -1,27 +1,24 @@
 import collections
 import collections.abc
 import functools
-import inspect
 import json
 import operator
-import sys
 import types
 import typing as tp
-from dataclasses import MISSING
-from dataclasses import fields as dataclass_fields
 from dataclasses import is_dataclass
 from functools import partial
 
 import attrs
-from attrs import NOTHING, Attribute, define
-from attrs import fields as attrs_fields
+from attrs import define
 from attrs import has as attrs_has
 
 from ..hints import resolve_types
 from ..memoized_property import memoized_property
 from ..not_specified import NotSpecifiedMeta
 from ..standard import builtin_types, union_types
+from .fields import Field, fields_from_attrs, fields_from_class, fields_from_dataclasses
 from .instance_check import InstanceCheck, create_checkable
+from .score import Score
 
 if tp.TYPE_CHECKING:
     from ..annotations import AdjustableCreator, AdjustableMeta
@@ -45,321 +42,6 @@ def _get_generic_super() -> type:
 
 
 generic_super = _get_generic_super()
-
-
-@define(order=True)
-class ScoreOrigin:
-    # The order of the fields matter
-    custom: bool = attrs.field(init=False)
-    package: str
-    module: str
-    name: str
-
-    def __attrs_post_init__(self) -> None:
-        self.custom = self.module != "builtins"
-
-    @classmethod
-    def create(self, typ: type) -> "ScoreOrigin":
-        return ScoreOrigin(
-            name=typ.__name__, module=typ.__module__, package=getattr(typ, "__package__", "")
-        )
-
-    def for_display(self, indent="") -> str:
-        def with_space(o: object) -> str:
-            s = str(o)
-
-            if s:
-                return f" {s}"
-            else:
-                return ""
-
-        lines = [
-            f"custom:{with_space(self.custom)}",
-            f"name:{with_space(self.name)}",
-            f"module:{with_space(self.module)}",
-            f"package:{with_space(self.package)}",
-        ]
-        return "\n".join(f"{indent}{line}" for line in lines)
-
-
-@define(order=True)
-class Score:
-    # The order of the fields matter
-    annotated_union: tuple["Score", ...] = attrs.field(init=False)
-    union_optional: bool = attrs.field(init=False)
-    union_length: int = attrs.field(init=False)
-    union: tuple["Score", ...]
-    annotated: bool
-    custom: bool = attrs.field(init=False)
-    optional: bool
-    mro_length: int = attrs.field(init=False)
-    typevars_length: int = attrs.field(init=False)
-    typevars_filled: tuple[bool, ...]
-    typevars: tuple["Score", ...]
-    origin_mro: tuple[ScoreOrigin, ...]
-
-    @classmethod
-    def create(cls, typ: "Type") -> "Score":
-        return cls(
-            union=tuple(ut.score for ut in typ.nonoptional_union_types),
-            typevars=tuple(tv.score for tv in typ.mro.all_vars),
-            typevars_filled=tuple(tv is not Type.Missing for tv in typ.mro.all_vars),
-            optional=typ.optional,
-            annotated=typ.is_annotated,
-            origin_mro=tuple(ScoreOrigin.create(t) for t in typ.origin.__mro__),
-        )
-
-    def __attrs_post_init__(self) -> None:
-        self.custom = False if not self.origin_mro else self.origin_mro[0].custom
-        self.union_length = len(self.union)
-        self.union_optional = bool(self.union) and self.optional
-        self.mro_length = len(self.origin_mro)
-        self.typevars_length = len(self.typevars)
-
-        if self.annotated and self.union:
-            self.annotated_union = self.union
-            self.union = ()
-        else:
-            self.annotated_union = ()
-
-    def for_display(self, indent="  ") -> str:
-        lines: list[str] = []
-
-        class WithDisplay(tp.Protocol):
-            def for_display(self, indent="") -> str:
-                ...
-
-        def extend(displayable: WithDisplay, extra: tp.Callable[[int], str]) -> None:
-            for i, line in enumerate(displayable.for_display(indent=indent).split("\n")):
-                lines.append(f"{extra(i)}{line}")
-
-        if self.annotated_union:
-            lines.append("✓ Annotated Union:")
-            for score in self.union or self.annotated_union:
-                extend(score, lambda i: "  *" if i == 0 else "   ")
-
-        if self.union_optional:
-            lines.append("✓ Union optional")
-        else:
-            lines.append("x Union optional")
-
-        if self.union_length:
-            lines.append(f"{self.union_length} Union length")
-
-        if self.union:
-            lines.append("✓ Union:")
-            for score in self.union or self.annotated_union:
-                extend(score, lambda i: "  *" if i == 0 else "   ")
-
-        if not self.annotated_union and not self.union:
-            lines.append("x Union")
-
-        if self.annotated:
-            lines.append("✓ Annotated")
-        else:
-            lines.append("x Annotated")
-
-        lines.append(f"{self.typevars_length} typevars {self.typevars_filled}")
-
-        if self.typevars:
-            lines.append("✓ Typevars:")
-            for score in self.typevars:
-                extend(score, lambda i: "  *" if i == 0 else "   ")
-        else:
-            lines.append("x Typevars")
-
-        if self.optional:
-            lines.append("✓ Optional")
-        else:
-            lines.append("x Optional")
-
-        lines.append(f"{self.mro_length} MRO length")
-
-        if self.origin_mro:
-            lines.append("✓ Origin MRO:")
-            for origin in self.origin_mro:
-                extend(origin, lambda i: "  *" if i == 0 else "   ")
-        else:
-            lines.append("x Origin MRO")
-
-        return "\n".join(f"{indent}{line}" for line in lines)
-
-
-@define
-class Default:
-    value: object | None
-
-    def __call__(self) -> object | None:
-        return self.value
-
-
-def kind_name(kind: int) -> str:
-    known = (
-        inspect.Parameter.KEYWORD_ONLY,
-        inspect.Parameter.VAR_POSITIONAL,
-        inspect.Parameter.POSITIONAL_OR_KEYWORD,
-        inspect.Parameter.POSITIONAL_ONLY,
-        inspect.Parameter.KEYWORD_ONLY,
-        inspect.Parameter.VAR_KEYWORD,
-    )
-
-    for k in known:
-        if k.value == kind:
-            return f"'{k.description}'"
-
-    return "<UNKNOWN_KIND>"
-
-
-@define
-class Field(tp.Generic[T]):
-    name: str
-    owner: object
-    disassembled_type: "Type[T]"
-    kind: int = attrs.field(default=inspect.Parameter.POSITIONAL_OR_KEYWORD.value, repr=kind_name)
-    default: tp.Callable[[], object | None] | None = attrs.field(default=None)
-    original_owner: object = attrs.field(default=attrs.Factory(lambda s: s.owner, takes_self=True))
-
-    def with_replaced_type(self, typ: "Type[U]") -> "Field[U]":
-        return Field[U](
-            name=self.name,
-            owner=self.owner,
-            disassembled_type=typ,
-            kind=self.kind,
-            default=self.default,
-            original_owner=self.original_owner,
-        )
-
-    @property
-    def type(self) -> object:
-        return self.disassembled_type.original
-
-    def clone(self) -> "Field[T]":
-        return self.with_replaced_type(self.disassembled_type)
-
-    @kind.validator
-    def check_kind(self, attribute: Attribute, value: object) -> None:
-        allowed = (
-            inspect.Parameter.KEYWORD_ONLY,
-            inspect.Parameter.VAR_POSITIONAL,
-            inspect.Parameter.POSITIONAL_OR_KEYWORD,
-            inspect.Parameter.POSITIONAL_ONLY,
-            inspect.Parameter.KEYWORD_ONLY,
-            inspect.Parameter.VAR_KEYWORD,
-        )
-        if value not in [a.value for a in allowed]:
-            raise ValueError(
-                f"Only allow parameter kinds. Got {value}, want one of {', '.join([f'{a.value} ({a.description})' for a in allowed])}"
-            )
-
-
-def fields_from_class(type_cache: "TypeCache", typ: type) -> tp.Sequence[Field]:
-    result: list[Field] = []
-    try:
-        signature = inspect.signature(typ)
-    except ValueError:
-        return result
-
-    for name, param in list(signature.parameters.items()):
-        field_type = param.annotation
-        if param.annotation is inspect.Parameter.empty:
-            field_type = object
-
-        if param.kind in (inspect.Parameter.VAR_KEYWORD, inspect.Parameter.VAR_POSITIONAL):
-            name = ""
-
-        dflt: tp.Callable[[], object | None] | None = None
-        if param.default is not inspect.Parameter.empty:
-            dflt = Default(param.default)
-        result.append(
-            Field(
-                name=name,
-                owner=typ,
-                default=dflt,
-                kind=param.kind.value,
-                disassembled_type=Type.create(field_type, cache=type_cache),
-            )
-        )
-
-    return result
-
-
-def fields_from_attrs(type_cache: "TypeCache", typ: type) -> tp.Sequence[Field]:
-    result: list[Field] = []
-    for field in attrs_fields(typ):
-        if not field.init:
-            continue
-
-        field_type = field.type
-        if field_type is None:
-            field_type = object
-
-        kind = inspect.Parameter.POSITIONAL_OR_KEYWORD.value
-        if field.kw_only:
-            kind = inspect.Parameter.KEYWORD_ONLY.value
-
-        dflt: tp.Callable[[], object | None] | None = None
-        if hasattr(field.default, "factory") and callable(field.default.factory):
-            if not field.default.takes_self:
-                dflt = field.default.factory
-
-        elif field.default is not NOTHING:
-            dflt = Default(field.default)
-
-        if sys.version_info >= (3, 11) and field.alias is not None:
-            name = field.alias
-        else:
-            name = field.name
-            if name.startswith("_"):
-                name = name[1:]
-
-        if name.startswith(f"{typ.__name__}_"):
-            name = name[len(f"{typ.__name__}_") + 1 :]
-
-        result.append(
-            Field(
-                name=name,
-                owner=typ,
-                default=dflt,
-                kind=kind,
-                disassembled_type=Type.create(field_type, cache=type_cache),
-            )
-        )
-
-    return result
-
-
-def fields_from_dataclasses(type_cache: "TypeCache", typ: type) -> tp.Sequence[Field]:
-    result: list[Field] = []
-    for field in dataclass_fields(typ):
-        if not field.init:
-            continue
-
-        field_type = field.type
-        if field_type is None:
-            field_type = object
-
-        kind = inspect.Parameter.POSITIONAL_OR_KEYWORD.value
-        if field.kw_only:
-            kind = inspect.Parameter.KEYWORD_ONLY.value
-
-        dflt: tp.Callable[[], object | None] | None = None
-        if field.default is not MISSING:
-            dflt = Default(field.default)
-
-        if field.default_factory is not MISSING:
-            dflt = field.default_factory
-
-        name = field.name
-        result.append(
-            Field(
-                name=name,
-                owner=typ,
-                default=dflt,
-                kind=kind,
-                disassembled_type=Type.create(field_type, cache=type_cache),
-            )
-        )
-    return result
 
 
 class IsAnnotated(tp.Protocol):
