+++ tests/disassemble/test_helpers.py
@@ -11,7 +11,6 @@ from attrs import define
 
 from strcs import Field, Type, TypeCache
 from strcs.disassemble.base import (
-    Default,
     IsAnnotated,
     extract_annotation,
     extract_optional,
@@ -19,6 +18,7 @@ from strcs.disassemble.base import (
     fields_from_class,
     fields_from_dataclasses,
 )
+from strcs.disassemble.fields import Default
 
 
 class Disassembler:
